app1 bugfix
Line 15: changed to keychain_winners = raffle_tickets[bottom_80_percent:] so the keychain winners are taken from the last 20% of the shuffled tickets (not from the eliminated front slice).

Line 16: changed to top_3_winners = raffle_tickets[-3:] so the top-3 winners are selected from the end of the shuffled list (the winners pool) instead of from the start.